+++ packages/entities/entities-gateway-services/src/components/GatewayServiceList.vue
@@ -101,6 +101,13 @@
               />
             </div>
           </template>
+
+          <template
+            v-if="config?.isControlPlaneGroup"
+            #message
+          >
+            {{ t('gateway_services.empty_state_v2.group') }}
+          </template>
         </EntityEmptyState>
       </template>