+++ packages/entities/entities-plugins/src/composables/useSchemas.ts
@@ -5,7 +5,6 @@ import { PLUGIN_METADATA } from '../definitions/metadata'
 import { aiPromptDecoratorSchema } from '../definitions/schemas/AIPromptDecorator'
 import { aiPromptTemplateSchema } from '../definitions/schemas/AIPromptTemplate'
 import { aiProxyAdvancedSchema } from '../definitions/schemas/AIProxyAdvanced'
-import { aiRateLimitingAdvancedSchema } from '../definitions/schemas/AIRateLimitingAdvanced'
 import { applicationRegistrationSchema } from '../definitions/schemas/ApplicationRegistration'
 import { ArrayInputFieldSchema } from '../definitions/schemas/ArrayInputFieldSchema'
 import { dataDogSchema } from '../definitions/schemas/Datadog'
@@ -142,10 +141,6 @@ export const useSchemas = (options?: UseSchemasOptions) => {
 
     'ai-proxy-advanced': {
       ...aiProxyAdvancedSchema,
-    },
-
-    'ai-rate-limiting-advanced': {
-      ...aiRateLimitingAdvancedSchema,
     },
 
     'vault-auth': {
@@ -288,11 +283,10 @@ export const useSchemas = (options?: UseSchemasOptions) => {
        * - Rendering a form for a plugin credential
        */
 
-      // Assume the fields are sorted, unless they have an `order` property
-      if (metadata?.useLegacyForm) {
+      // We group redis fields separately only when this plugin supports redis partial and redisPartial is enabled
+      if (metadata?.useLegacyForm && options?.enableRedisPartial && currentSchema._supported_redis_partial_type) {
         for (const field of formSchema.fields!) {
-          // We group redis fields separately only when this plugin supports redis partial and redisPartial is enabled
-          if (options?.enableRedisPartial && currentSchema._supported_redis_partial_type && isRedisField(field)) {
+          if (isRedisField(field)) {
             redisFields.push(field)
             continue
           }
@@ -309,7 +303,7 @@ export const useSchemas = (options?: UseSchemasOptions) => {
         })
       }
 
-
+      // Assume the fields are sorted, unless they have an `order` property
       formSchema.fields!.sort((a: Record<string, any>, b: Record<string, any>) => {
         a.order = a.order || 0
         b.order = b.order || 0