+++ packages/core/expressions/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.12...@kong-ui-public/expressions@0.9.13) (2025-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.9.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.11...@kong-ui-public/expressions@0.9.12) (2025-03-04)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.9.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.10...@kong-ui-public/expressions@0.9.11) (2025-02-28)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.9.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.9...@kong-ui-public/expressions@0.9.10) (2025-02-28)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.9.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.8...@kong-ui-public/expressions@0.9.9) (2025-02-28)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
 ## [0.9.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.9.7...@kong-ui-public/expressions@0.9.8) (2025-02-27)
 
 **Note:** Version bump only for package @kong-ui-public/expressions