+++ packages/core/tracing/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/tracing@0.9.4...@kong-ui-public/tracing@0.9.5) (2025-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/tracing
+
+
+
+
+
+## [0.9.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/tracing@0.9.3...@kong-ui-public/tracing@0.9.4) (2025-03-04)
+
+**Note:** Version bump only for package @kong-ui-public/tracing
+
+
+
+
+
 ## [0.9.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/tracing@0.9.2...@kong-ui-public/tracing@0.9.3) (2025-02-27)
 
 **Note:** Version bump only for package @kong-ui-public/tracing