+++ packages/entities/entities-consumer-credentials/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.53](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@3.4.52...@kong-ui-public/entities-consumer-credentials@3.4.53) (2025-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials
+
+
+
+
+
+## [3.4.52](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@3.4.51...@kong-ui-public/entities-consumer-credentials@3.4.52) (2025-03-04)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials
+
+
+
+
+
 ## [3.4.51](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@3.4.50...@kong-ui-public/entities-consumer-credentials@3.4.51) (2025-02-27)
 
 **Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials