+++ packages/entities/entities-key-sets/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-key-sets@3.7.4...@kong-ui-public/entities-key-sets@3.7.5) (2025-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/entities-key-sets
+
+
+
+
+
+## [3.7.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-key-sets@3.7.3...@kong-ui-public/entities-key-sets@3.7.4) (2025-03-04)
+
+**Note:** Version bump only for package @kong-ui-public/entities-key-sets
+
+
+
+
+
 ## [3.7.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-key-sets@3.7.2...@kong-ui-public/entities-key-sets@3.7.3) (2025-02-27)
 
 **Note:** Version bump only for package @kong-ui-public/entities-key-sets