+++ packages/entities/entities-redis-configurations/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.1.9...@kong-ui-public/entities-redis-configurations@0.1.10) (2025-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.1.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.1.8...@kong-ui-public/entities-redis-configurations@0.1.9) (2025-03-05)
+
+
+### Bug Fixes
+
+* **redis:** add password mask ([#2003](https://github.com/Kong/public-ui-components/issues/2003)) ([4beb036](https://github.com/Kong/public-ui-components/commit/4beb036e8d4d91f0bc85ca9283dc62e631d02460))
+
+
+
+
+
+## [0.1.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.1.7...@kong-ui-public/entities-redis-configurations@0.1.8) (2025-03-04)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.1.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.1.6...@kong-ui-public/entities-redis-configurations@0.1.7) (2025-03-03)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.1.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.1.5...@kong-ui-public/entities-redis-configurations@0.1.6) (2025-02-28)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.1.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.1.4...@kong-ui-public/entities-redis-configurations@0.1.5) (2025-02-28)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.1.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.1.3...@kong-ui-public/entities-redis-configurations@0.1.4) (2025-02-28)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.1.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.1.2...@kong-ui-public/entities-redis-configurations@0.1.3) (2025-02-28)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
 ## [0.1.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.1.1...@kong-ui-public/entities-redis-configurations@0.1.2) (2025-02-27)
 
 **Note:** Version bump only for package @kong-ui-public/entities-redis-configurations