+++ packages/entities/entities-shared/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.19.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.19.1...@kong-ui-public/entities-shared@3.19.2) (2025-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
+## [3.19.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.19.0...@kong-ui-public/entities-shared@3.19.1) (2025-03-04)
+
+
+### Bug Fixes
+
+* entity link truncate on max width 90% ([#2000](https://github.com/Kong/public-ui-components/issues/2000)) ([bba8c6f](https://github.com/Kong/public-ui-components/commit/bba8c6fb8385a4d5fecd03c9a09db7cdd096bce7))
+
+
+
+
+
 # [3.19.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.18.0...@kong-ui-public/entities-shared@3.19.0) (2025-02-27)